Moved findClassByName from cc.core to cc.staticanalyzer plugin
