+++ plugins/de.cognicrypt.core/META-INF/MANIFEST.MF
@@ -13,8 +13,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.jdt.core,
  com.google.guava,
  org.eclipse.ui.forms,
- org.eclipse.m2e.maven.runtime,
- org.jetbrains.kotlin.core;resolution:=optional;visibility:=reexport
+ org.eclipse.m2e.maven.runtime
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Automatic-Module-Name: de.cognicrypt.core
